# Fix headless being limited to 1000hz